Refatora aba de úteis para componente dedicado
